+++ ci/licenses_golden/licenses_flutter
@@ -34,12 +34,23 @@ FILE: ../../../flutter/common/task_runners.cc
 FILE: ../../../flutter/common/task_runners.h
 FILE: ../../../flutter/display_list/display_list.cc
 FILE: ../../../flutter/display_list/display_list.h
-FILE: ../../../flutter/display_list/display_list_canvas.cc
-FILE: ../../../flutter/display_list/display_list_canvas.h
+FILE: ../../../flutter/display_list/display_list_builder.cc
+FILE: ../../../flutter/display_list/display_list_builder.h
+FILE: ../../../flutter/display_list/display_list_canvas_dispatcher.cc
+FILE: ../../../flutter/display_list/display_list_canvas_dispatcher.h
+FILE: ../../../flutter/display_list/display_list_canvas_recorder.cc
+FILE: ../../../flutter/display_list/display_list_canvas_recorder.h
 FILE: ../../../flutter/display_list/display_list_canvas_unittests.cc
+FILE: ../../../flutter/display_list/display_list_dispatcher.cc
+FILE: ../../../flutter/display_list/display_list_dispatcher.h
+FILE: ../../../flutter/display_list/display_list_flags.cc
+FILE: ../../../flutter/display_list/display_list_flags.h
+FILE: ../../../flutter/display_list/display_list_ops.cc
+FILE: ../../../flutter/display_list/display_list_ops.h
 FILE: ../../../flutter/display_list/display_list_unittests.cc
 FILE: ../../../flutter/display_list/display_list_utils.cc
 FILE: ../../../flutter/display_list/display_list_utils.h
+FILE: ../../../flutter/display_list/types.h
 FILE: ../../../flutter/flow/compositor_context.cc
 FILE: ../../../flutter/flow/compositor_context.h
 FILE: ../../../flutter/flow/diff_context.cc
