+++ display_list/display_list_flags.cc
@@ -1,0 +1,11 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "flutter/display_list/display_list_flags.h"
+
+namespace flutter {
+
+// Just exists to ensure that the header can be cleanly imported.
+
+}  // namespace flutter
