+++ flow/testing/diff_context_test.cc
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 #include "diff_context_test.h"
+#include "flutter/display_list/display_list_builder.h"
 
 namespace flutter {
 namespace testing {
