+++ lib/ui/painting/paint.h
@@ -6,6 +6,7 @@
 #define FLUTTER_LIB_UI_PAINTING_PAINT_H_
 
 #include "flutter/display_list/display_list.h"
+#include "flutter/display_list/display_list_flags.h"
 #include "third_party/skia/include/core/SkPaint.h"
 #include "third_party/tonic/converter/dart_converter.h"
